% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaDificil.R
\name{mediaDificil}
\alias{mediaDificil}
\title{A primeira funcao q fiz.}
\usage{
mediaDificil(x)
}
\arguments{
\item{x}{vetor numerico sobre o qual se quer calcular a media.}
}
\description{
A primeira funcao q fiz.
}
